Rename message senders
